Highlighter#resetCheckedState: modify locking
Use addPendingAction to run the reset in the highlighter thread instead
of synchronizing on the HighlighterEditorTracker. This seems more
appropriate to me since the reset has to clear the checked editors list
and also reset each checker, and both the checkers and the editors are
accessed from (and primarily locked by) the highlighter thread.