MPS-23899 Fix missing icon for Pin Tab button
The problem was caused by a change in IDEA. Now the icon for the
standard PinToolwindowTabAction is only set if the action is in a
"toolbar place". Our actions were mostly put into toolbars with
ActionPlaces.UNKNOWN, now we set a proper place for toolbars that
contain the Pin Tab action.

Also, PinToolwindowTabAction was deprecated, so switch to its
replacement PinActiveTabAction.