Correcting undo/redo behavior for commands creating/removing root nodes.
In the past such commands has no affectedDocuments associated, so cannot be executed in a context of other editors, touched by the same command. Effectively this leads us to the inconsistent undo stacks for such editors.

Eg: Command, modifying current node and creating another root one. This command will be "global". If there are no documents associated with it, then on pressing undo in the same editor (right after running the command) this command will not be executed. Instead, previous command will be executed probably throwing some exceptions and breaking undo consistency for the current editor...