MPS-24504 cmd+alt+up/down is unusable
When switch to next/prev occurence in UsagesView, user want UsagesView to stay focused and apply action as many times as he needs.
On other side for Node it is better to show them in Editor, while for modules and models there is no other way then to show them in ProjectPane.