Use SimpleToolWindowPanel in tool windows to get nicer borders
Affected tool windows:
- Hierarchy View (and Class Hierarchy View)
- Messages View
- Console
- Inspector

The tool window area has its own border so the component and the tool
bar don't need their own border. The content components often had a
border around them, thus the border was double-width in some places and
single-width in others.

SimpleToolWindowPanel lays out a toolbar and a content component with a
thin border between them, so when the border around the content is
removed, only single-width border remains everywhere.